css/css-shadow-parts/inner-host.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Inner Host</title>
    <meta href="fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="http://www.w3.org/TR/css-variables-1/#syntax" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <style>#c-e-green::part(partp) { color: #00ff00; }</style>
    <script type="text/javascript">installCustomElement("custom-element-inner", "custom-element-inner-template");</script>
    <template id="custom-element-inner-template">
      <style>span { color: #0000ff; }</style>
      <span id="blue_part" part="partp">This text</span>
    </template>
    <script type="text/javascript">installCustomElement("custom-element-outer", "custom-element-outer-template");</script>
    <template id="custom-element-outer-template">
      <style>span { color: #ff0000; }</style>
      <span id="green_part" part="partp">This text</span>
      The following text should be blue:
      <custom-element-inner id="c-e-blue"></custom-element-inner>
    </template>
    The following text should be green:
    <custom-element-outer id="c-e-green"></custom-element-outer>
    <script type="text/javascript">
      "use strict";
      test(function() {
        var el = getElementByShadowIds(document, ["c-e-green", "green_part"]);
        assert_equals(window.getComputedStyle(el).color, "rgb(0, 255, 0)");
      }, "Part in outer host is styled by document style sheet");
      test(function() {
        var el = getElementByShadowIds(document, ["c-e-green", "c-e-blue", "blue_part"]);
        assert_equals(window.getComputedStyle(el).color, "rgb(0, 0, 255)");
      }, "Part in inner host is not styled by document style sheet");
    </script>
  </body>
</html>
