LayoutTests/imported/w3c/web-platform-tests/css/css-shadow-parts/inner-host-expected.txt

The following text should be green: 

FAIL Part in outer host is styled by document style sheet assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
PASS Part in inner host is not styled by document style sheet 
